+++ lib/views/vote/buttons/base/vote_button.dart
