+++ lib/views/vote/buttons/down_vote_button.dart
@@ -2,11 +2,11 @@ import 'package:flutter/material.dart';
 import 'package:pr0gramm/constants/vote_constants.dart';
 import 'package:pr0gramm/entities/enums/vote_button_type.dart';
 import 'package:pr0gramm/services/vote_animation_service.dart';
+import 'package:pr0gramm/views/vote/mixins/vote_button_color_animation.dart';
+import 'package:pr0gramm/views/vote/mixins/vote_button_rotate_animation.dart';
 import 'package:pr0gramm/views/vote/sized_vote_button.dart';
-import 'package:pr0gramm/views/vote/vote_button.dart';
-import 'package:pr0gramm/views/vote/vote_button_animation_integration.dart';
-import 'package:pr0gramm/views/vote/vote_button_color_animation.dart';
-import 'package:pr0gramm/views/vote/vote_button_rotate_animation.dart';
+import 'package:pr0gramm/views/vote/buttons/base/vote_button.dart';
+import 'package:pr0gramm/views/vote/mixins/vote_button_animation_integration.dart';
 
 class DownVoteButton extends VoteButton {
   DownVoteButton({
@@ -34,9 +34,9 @@ class _DownVoteButtonState extends State<DownVoteButton>
     with
         TickerProviderStateMixin,
         SizedVoteButton,
-        VoteButtonAnimationIntegration<DownVoteButton>,
-        VoteButtonColorAnimation<DownVoteButton>,
-        VoteButtonRotateAnimation<DownVoteButton> {
+        VoteButtonAnimationIntegration,
+        VoteButtonColorAnimation,
+        VoteButtonRotateAnimation {
   @override
   void initState() {
     super.initState();
