+++ lib/views/vote/buttons/favorite_vote_button.dart
@@ -2,10 +2,10 @@ import 'package:flutter/material.dart';
 import 'package:pr0gramm/constants/vote_constants.dart';
 import 'package:pr0gramm/entities/enums/vote_button_type.dart';
 import 'package:pr0gramm/services/vote_animation_service.dart';
+import 'package:pr0gramm/views/vote/mixins/vote_button_color_animation.dart';
 import 'package:pr0gramm/views/vote/sized_vote_button.dart';
-import 'package:pr0gramm/views/vote/vote_button.dart';
-import 'package:pr0gramm/views/vote/vote_button_animation_integration.dart';
-import 'package:pr0gramm/views/vote/vote_button_color_animation.dart';
+import 'package:pr0gramm/views/vote/buttons/base/vote_button.dart';
+import 'package:pr0gramm/views/vote/mixins/vote_button_animation_integration.dart';
 
 class FavoriteVoteButton extends VoteButton {
   FavoriteVoteButton({
@@ -33,8 +33,8 @@ class _FavoriteVoteButtonState extends State<FavoriteVoteButton>
     with
         TickerProviderStateMixin,
         SizedVoteButton,
-        VoteButtonAnimationIntegration<FavoriteVoteButton>,
-        VoteButtonColorAnimation<FavoriteVoteButton> {
+        VoteButtonAnimationIntegration,
+        VoteButtonColorAnimation {
   AnimationController _controller;
   CurvedAnimation _morphCurve;
   Animation<double> _morphProgress;
