+++ lib/views/vote/comment_vote.dart
@@ -3,9 +3,10 @@ import 'package:pr0gramm/api/dtos/comment/comment.dart';
 import 'package:pr0gramm/entities/enums/vote.dart';
 import 'package:pr0gramm/services/vote_animation_service.dart';
 import 'package:pr0gramm/services/vote_service.dart';
-import 'package:pr0gramm/views/vote/down_vote_button.dart';
-import 'package:pr0gramm/views/vote/up_vote_button.dart';
+import 'package:pr0gramm/views/vote/buttons/up_vote_button.dart';
 import 'package:pr0gramm/widgets/global_inherited.dart';
+
+import 'buttons/down_vote_button.dart';
 
 class CommentVote extends StatelessWidget {
   final Comment comment;
