+++ lib/views/vote/mixins/vote_button_animation_integration.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:pr0gramm/services/vote_animation_service.dart';
-import 'package:pr0gramm/views/vote/vote_button.dart';
+import 'package:pr0gramm/views/vote/buttons/base/vote_button.dart';
 
 mixin VoteButtonAnimationIntegration<T extends VoteButton>
     on State<T>, TickerProvider {
